# src/apiDocs.json
{
   "openapi": "3.0.0",
   "components": {
      "securitySchemes": {
         "cookieAuth": {
            "type": "apiKey",
            "in": "cookie",
            "name": "token"
         }
      }
   },
   "info": {
      "title": "API Boilerplate",
      "description": "Daftar api Boilerplate PT. XYZ",
      "version": "1.0.0"
   },
...
   "paths": {
...
      "/api/users": {
         "get": {
            "tags": ["Users"],
            "summary": "Get all users",
            "description": "API untuk mengambil data users",
            "produces": "application/json",
            "parameters": [
               {
                  "in": "query",
                  "name": "id",
                  "schema": {
                     "type": "string"
                  },
                  "description": "Find user by id"
               },
               {
                  "in": "query",
                  "name": "name",
                  "schema": {
                     "type": "string"
                  },
                  "description": "Find user by name"
               },
               {
                  "in": "query",
                  "name": "email",
                  "schema": {
                     "type": "string"
                  },
                  "description": "Find user by email"
               },
               {
                  "in": "query",
                  "name": "isAdmin",
                  "schema": {
                     "type": "boolean"
                  },
                  "description": "Find user by isAdmin"
               },
               {
                  "in": "query",
                  "name": "limit",
                  "schema": {
                     "type": "integer"
                  },
                  "description": "The number limit of users. default 100"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "schema": {
                     "$ref": "#/definitions/Users"
                  }
               }
            }
         },
         "post": {
            "tags": ["Users"],
            "summary": "Create user",
            "description": "This can only be done by the logged in user.",
            "operationId": "createUser",
            "consumes": ["multipart/form-data"],
            "requestBody": {
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "$ref": "#/definitions/Users"
                     }
                  }
               }
            },
            "responses": {
               "default": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/api/users/{id}": {
         "put": {
            "tags": ["Users"],
            "summary": "Update user",
            "description": "This can only be done by the logged in user.",
            "operationId": "updateUser",
            "consumes": ["multipart/form-data"],
            "parameters": [
               {
                  "in": "path",
                  "name": "id",
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "requestBody": {
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "$ref": "#/definitions/Users"
                     }
                  }
               }
            },
            "responses": {
               "default": {
                  "description": "successful operation"
               }
            }
         },
         "delete": {
            "tags": ["Users"],
            "summary": "Delete user",
            "description": "This can only be done by the logged in user.",
            "operationId": "delete",
            "security": [
               {
                  "cookieAuth": []
               }
            ],
            "parameters": [
               {
                  "in": "path",
                  "name": "id",
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "responses": {
               "default": {
                  "description": "successful operation"
               }
            }
         }
      }
   },
   "definitions": {
      "Users": {
         "type": "object",
         "properties": {
            "name": {
               "type": "string"
            },
            "photo": {
               "type": "string",
               "format": "binary"
            },
            "email": {
               "type": "string"
            },
            "password": {
               "type": "string"
            },
            "isAdmin": {
               "type": "boolean",
               "description": "only admin can change this field"
            },
            "gender": {
               "type": "string",
               "enum": ["L", "P"]
            }
         },
         "xml": {
            "name": "Users"
         }
      }
   }
}